Moved all types to own file
